Prefer dimens. of current slide instead of first
This fixes a problem in that if the user deep-links to a slide that is not
the first and the first has display:none, slideHeight and slideWidth would
be 0 and break the bespoke-scale logic.

The motivation to be able to hide some slides is that some themes (e.g.,
bespoke-theme-fancy) might want to hide non-current slides to improve
performance.